server/deviations/serializers/grade_range_serializer.py
from rest_framework import serializers
from ..models import GradeRange


class GradeRangeSerializer(serializers.ModelSerializer):
    class Meta:
        model = GradeRange
        fields = "__all__"
